requirements.txt
fastapi[standard]
uvicorn
python-multipart
openai
python-dotenv
google-genai
elevenlabs
moviepy
google-generativeai 
apscheduler==3.11.2